+++ undine/client/database/client_base.py
@@ -7,3 +7,9 @@ class ClientBase:
 
     def mission_info(self, **kwargs):
         raise VirtualMethodException(ClientBase, 'mission_info')
+
+    def task_list(self, **kwargs):
+        raise VirtualMethodException(ClientBase, 'task_list')
+
+    def task_info(self, tid):
+        raise VirtualMethodException(ClientBase, 'task_info')